+++ db/schema/index.js
@@ -2,3 +2,5 @@ export * from './subscribers.js';
 export * from './parutions.js';
 export * from './annonces.js';
 export * from './envois.js';
+export * from './conversations.js';
+export * from './bot_responses.js';